requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.62.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
pip==24.0
psycopg2-binary
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
